requirements.txt
python-dotenv>=1.0.0
python-dateutil>=2.9.0

Flask>=3.0.0

langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langchain-groq==0.3.8
langchain-huggingface==0.1.2
groq>=0.31.1,<1.0.0

faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
